Add support for no_run info string
This commit adds support for the `no_run` info string
which works similar to how it works for rustdoc (code is
compiled but not executed).  Blocks marked as `no_run`
result in a generated function that is dead code (does
not have `test` attribute applied).

This is pretty useful for examples in documentation which
have side effects and would frequently fail or do
undesirable things in a test environment.

I also updated the documentation to make getting details on
each info string a bit simpler.

Signed-off-by: Paul Osborne <osbpau@gmail.com>