feat: Put the test code inline into the test module
This fixes two major issues with skeptic's current implementation at the
cost of losing compile-fail tests. For many crates I expect this to be a
worthwhile tradeoff. It is possible to restore compile-tests but they
would be forced to use the same shelling out approach as before this PR
so they would suffer from the same problem. As such I prefer to keep
them out instead.

No longer shells out to rustc and attempts to emulate cargo's dependency
resolution.

Fixes #18

Tests are now in a normal test module letting cargo cache and reuse
compilation artifacts as normal.

Fixes #8

BREAKING CHANGE

- Handlebars templates are now used instead of `format!`
  Since there is no intermediate step we need a way to dynamically
  template. Handlebars is well known and well supported.

- Compile tests that expect panics are not supported
  Since code is generated inline, these tests would cause test
  compilation to fail